Plugin support for custom sidebar (#327)

+++ src/ComicReaderUWP/Common/Plugins/PluginManager.cs
@@ -7,6 +7,8 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 using System.Text.RegularExpressions;
 using System.Threading;
 
@@ -90,40 +92,7 @@ internal partial class PluginManager
 
             if (plugins.Count > 0)
             {
-                foreach (PluginContext plugin in plugins)
-                {
-                    foreach (string assembly in plugin.Plugin.SharedAssemblies)
-                    {
-                        loadContext.AssemblyLoader.AddSharedAssemblyName(assembly);
-                    }
-                }
-
-                List<IXamlMetadataProvider> xamlMetaProviders = [.. loadContext.XamlMetadataProviders];
-
-                foreach (string assemblyName in loadContext.AssemblyLoader.SharedAssemblies)
-                {
-                    if (!loadContext.Assemblies.TryGetValue(assemblyName, out string? assemblyPath))
-                    {
-                        Logger.E(TAG, $"Failed to find shared assembly '{assemblyName}' for plugin '{pluginFile}'");
-                        continue;
-                    }
-
-                    Assembly assembly;
-                    try
-                    {
-                        assembly = Assembly.LoadFrom(assemblyPath);
-                    }
-                    catch (Exception ex)
-                    {
-                        Logger.E(TAG, $"Failed to load shared assembly '{assemblyName}' for plugin '{pluginFile}'", ex);
-                        continue;
-                    }
-
-                    xamlMetaProviders.AddRange(CreateInstancesFromAssembly<IXamlMetadataProvider>(assembly));
-                    Logger.I(TAG, $"Loaded shared assembly '{assemblyName}' for plugin '{pluginFile}'");
-                }
-
-                PluginXamlMetadataProvider.AddProviders(xamlMetaProviders);
+                PluginXamlMetadataProvider.AddProviders(loadContext.XamlMetadataProviders);
 
                 foreach (PluginContext plugin in plugins)
                 {
@@ -286,7 +255,6 @@ internal partial class PluginManager
         string extension = Path.GetExtension(pluginFile).ToLowerInvariant();
         return extension switch
         {
-            ".dll" => LoadDllPlugin(pluginFile),
             ".zip" => LoadZipPlugin(pluginFile),
             _ => null,
         };
@@ -296,28 +264,48 @@ internal partial class PluginManager
     {
         string pluginFileName = Path.GetFileNameWithoutExtension(pluginFile);
         string extractDir = Path.Combine(StorageLocation.TemporaryFolderPath, "plugins", pluginFileName);
-        try
+
+        string sourceSignature = FileUtils.GetFileSignature(pluginFile);
+        if (string.IsNullOrEmpty(sourceSignature))
         {
-            Directory.Delete(extractDir, true);
-        }
-        catch (DirectoryNotFoundException)
-        {
-            // Ignore
-        }
-        catch (Exception ex)
-        {
-            Logger.F(TAG, ex);
+            Logger.E(TAG, $"Failed to calculate signature for plugin '{pluginFile}'");
             return null;
         }
 
-        try
+        string signatureFile = Path.Combine(extractDir, "PluginSignature.txt");
+        if (!SignatureMatched(sourceSignature, signatureFile))
         {
-            Directory.CreateDirectory(extractDir);
-            System.IO.Compression.ZipFile.ExtractToDirectory(pluginFile, extractDir);
+            try
+            {
+                Directory.Delete(extractDir, true);
+            }
+            catch (DirectoryNotFoundException)
+            {
+            }
+            catch (Exception ex)
+            {
+                Logger.F(TAG, ex);
+                return null;
+            }
+
+            try
+            {
+                Directory.CreateDirectory(extractDir);
+                System.IO.Compression.ZipFile.ExtractToDirectory(pluginFile, extractDir);
+                File.WriteAllText(signatureFile, sourceSignature);
+            }
+            catch (Exception ex)
+            {
+                Logger.F(TAG, ex);
+                return null;
+            }
         }
-        catch (Exception ex)
+
+        string metaFile = Path.Combine(extractDir, "PluginMeta.json");
+        PluginMeta? meta = LoadPluginMeta(metaFile);
+        if (meta is null)
         {
-            Logger.F(TAG, ex);
+            Logger.E(TAG, $"Failed to load PluginMeta.json for plugin '{pluginFile}'");
             return null;
         }
 
@@ -343,30 +331,38 @@ internal partial class PluginManager
             return null;
         }
 
-        string mainAssemblyPath = assemblies[mainAssemblyName];
-        PluginFileLoadContext? loadDllResult = LoadDllPlugin(mainAssemblyPath);
-        if (loadDllResult is null)
+        // Load shared assemblies
+        List<IXamlMetadataProvider> xamlMetaProviders = [];
+        foreach (string assemblyName in meta.SharedAssemblies)
         {
-            return null;
+            string assemblyPath = Path.Combine(extractDir, assemblyName + ".dll");
+            Assembly assembly;
+            try
+            {
+                assembly = Assembly.LoadFrom(assemblyPath);
+            }
+            catch (Exception ex)
+            {
+                Logger.E(TAG, $"Failed to load shared assembly '{assemblyName}' for plugin '{pluginFile}'", ex);
+                continue;
+            }
+
+            xamlMetaProviders.AddRange(CreateInstancesFromAssembly<IXamlMetadataProvider>(assembly));
+            Logger.I(TAG, $"Loaded shared assembly '{assemblyName}' for plugin '{pluginFile}'");
         }
 
-        return new()
+        // Load main assembly
+        string mainAssemblyPath = assemblies[mainAssemblyName];
+        PluginAssemblyLoader loadContext = new(mainAssemblyPath);
+        foreach (string assemblyName in meta.SharedAssemblies)
         {
-            AssemblyLoader = loadDllResult.AssemblyLoader,
-            Plugins = loadDllResult.Plugins,
-            XamlMetadataProviders = loadDllResult.XamlMetadataProviders,
-            ResourceFolderPath = extractDir,
-            Assemblies = assemblies,
-        };
-    }
+            loadContext.AddSharedAssemblyName(assemblyName);
+        }
 
-    private static PluginFileLoadContext? LoadDllPlugin(string pluginFile)
-    {
-        PluginAssemblyLoader loadContext = new(pluginFile);
-        Assembly assembly;
+        Assembly mainAssembly;
         try
         {
-            assembly = loadContext.LoadFromAssemblyPath(pluginFile);
+            mainAssembly = loadContext.LoadFromAssemblyPath(mainAssemblyPath);
         }
         catch (Exception ex)
         {
@@ -374,14 +370,70 @@ internal partial class PluginManager
             return null;
         }
 
-        List<IPlugin> plugins = CreateInstancesFromAssembly<IPlugin>(assembly);
-        List<IXamlMetadataProvider> xamlMetadataProviders = CreateInstancesFromAssembly<IXamlMetadataProvider>(assembly);
+        List<IPlugin> plugins = CreateInstancesFromAssembly<IPlugin>(mainAssembly);
+        xamlMetaProviders.AddRange(CreateInstancesFromAssembly<IXamlMetadataProvider>(mainAssembly));
+
         return new()
         {
             AssemblyLoader = loadContext,
             Plugins = plugins,
-            XamlMetadataProviders = xamlMetadataProviders,
-            Assemblies = new Dictionary<string, string> { { Path.GetFileNameWithoutExtension(pluginFile), pluginFile } },
+            XamlMetadataProviders = xamlMetaProviders,
+            ResourceFolderPath = extractDir,
+            Assemblies = assemblies,
+        };
+    }
+
+    private static bool SignatureMatched(string sourceSignature, string signatureFile)
+    {
+        string cacheSignature;
+        try
+        {
+            cacheSignature = File.ReadAllText(signatureFile);
+        }
+        catch (FileNotFoundException)
+        {
+            return false;
+        }
+        catch (DirectoryNotFoundException)
+        {
+            return false;
+        }
+        catch (Exception ex)
+        {
+            Logger.F(TAG, ex);
+            return false;
+        }
+
+        if (string.IsNullOrEmpty(cacheSignature))
+        {
+            return false;
+        }
+
+        return cacheSignature == sourceSignature;
+    }
+
+    private static PluginMeta? LoadPluginMeta(string filePath)
+    {
+        PluginMetaJsonModel? jsonModel;
+        try
+        {
+            string json = File.ReadAllText(filePath);
+            jsonModel = JsonSerializer.Deserialize<PluginMetaJsonModel>(json);
+        }
+        catch (Exception ex)
+        {
+            Logger.E(TAG, ex);
+            return null;
+        }
+
+        if (jsonModel is null)
+        {
+            return null;
+        }
+
+        return new()
+        {
+            SharedAssemblies = jsonModel.SharedAssemblies,
         };
     }
 
@@ -422,4 +474,15 @@ internal partial class PluginManager
 
     [GeneratedRegex(@"^[a-zA-Z0-9_]+$")]
     private static partial Regex PluginNameRegex();
+
+    private class PluginMeta
+    {
+        public List<string> SharedAssemblies { get; set; } = [];
+    }
+
+    private class PluginMetaJsonModel
+    {
+        [JsonPropertyName("SharedAssemblies")]
+        public List<string> SharedAssemblies { get; set; } = [];
+    }
 }
